requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
python-dotenv==0.21.0
flask-ldap3-login==0.9.18
